Struct grpcio::GrpcSlice

#[repr(C)]pub struct GrpcSlice(_);

A convenient rust wrapper for the type grpc_slice.

It’s expected that the slice should be initialized.

Implementations

impl GrpcSlice

pub fn len(&self) -> usize

Get the length of the data.

pub fn as_slice(&self) -> &[u8]

Returns a slice of inner buffer.

pub fn is_empty(&self) -> bool

pub fn from_static_slice(s: &'static [u8]) -> GrpcSlice

Creates a slice from static rust slice.

Same as From<&[u8]> but without copying the buffer.

pub fn from_static_str(s: &'static str) -> GrpcSlice

Creates a GrpcSlice from static rust str.

Same as from_str but without allocation.

pub fn is_inline(&self) -> bool

Checks whether the slice stores bytes inline.

pub unsafe fn realloc(&mut self, cap: usize) -> &mut [MaybeUninit<u8>]

Reallocates current slice with given capacity.

The length of returned slice is the exact same as given cap.

Safety

Caller is expected to initialize all available bytes to guarantee safety of this slice.

pub fn as_mut_ptr(&mut self) -> *mut grpc_slice